[api] Avoid circular dependencies
